Move CMIP6 processor into plugins/cmip6; add processing shim; format and lint pass
