+++ src/piddiplatsch/records/cmip6_record.py
@@ -7,8 +7,8 @@ from piddiplatsch.config import config
 from piddiplatsch.exceptions import LookupError
 from piddiplatsch.lookup.api import get_lookup
 from piddiplatsch.models.base import HostingNode
-from piddiplatsch.plugins.cmip6.model import CMIP6DatasetModel, CMIP6FileModel
 from piddiplatsch.monitoring import stats
+from piddiplatsch.plugins.cmip6.model import CMIP6DatasetModel, CMIP6FileModel
 from piddiplatsch.records.base import BaseCMIP6Record
 from piddiplatsch.utils.models import (
     asset_pid,
